requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
inflection==0.5.1
itsdangerous==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
